scripts/generate-ssh-host-keys/requirements.txt
pycryptodome
ruamel.yaml
